expo_backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
ultralytics>=8.0.0
easyocr>=1.7.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
